requirements.txt
certifi==2018.1.18
chardet==3.0.4
Django==2.0.3
djangorestframework==3.7.7
geopy==1.12.0
gunicorn==19.7.1
httpie==0.9.9
idna==2.6
numpy==1.14.2
pandas==0.22.0
Pygments==2.2.0
pymongo==3.6.1
python-dateutil==2.7.0
pytz==2018.3
requests==2.18.4
six==1.11.0
urllib3==1.22
